Add xp bonus from scrolls

+++ src/structs/recruits.rs
@@ -100,15 +100,13 @@ impl RecruitInventory {
         return multiplicator;
     }
 
-    #[allow(dead_code)]
-    // WIP - Not tested yet
-    pub fn get_experience_multiplicator_from_scroll_bonus(&self) -> f32 {
+    pub fn get_experience_multiplicator_from_scroll_bonus(&self) -> f64 {
         let mut multiplicator = 1.0;
 
         for scroll in &self.scrolls {
             for bonus in &scroll.bonus {
                 if let BonusEnum::Experience(value) = bonus {
-                    multiplicator += *value as f32 / 100.;
+                    multiplicator += *value as f64 / 100.;
                 }
             }
         }
